requirements.txt
pandas==2.0.0
numpy==1.24.0
ta==0.10.2
tqdm==4.65.0
pyyaml==6.0.1
numba==0.57.1